Added information, created Transform AssetData
Something I noticed: It is entirely possible that ClassIDs and TypeIDs will only work within the file that they are used in. It is possible that in other files they are mapped differently. For this reason, we should instead use the hashes for scripts instead of the PathID, ClassID, TypeIDs.